More datatype cleanup

+++ Emulator/CPU/CPU.h
@@ -308,8 +308,8 @@ public:
     u16 getPC0() const { return reg.pc0; }
     
     void jumpToAddress(u16 addr) { reg.pc0 = reg.pc = addr; next = fetch; }
-    void setPCL(u8 lo) { reg.pc = (reg.pc & 0xff00) | lo; }
-    void setPCH(u8 hi) { reg.pc = (reg.pc & 0x00ff) | ((u16)hi << 8); }
+    void setPCL(u8 lo) { reg.pc = (u16)((reg.pc & 0xff00) | lo); }
+    void setPCH(u8 hi) { reg.pc = (u16)((reg.pc & 0x00ff) | hi << 8); }
     void incPC(u8 offset = 1) { reg.pc += offset; }
     void incPCL(u8 offset = 1) { setPCL(LO_BYTE(reg.pc) + offset); }
     void incPCH(u8 offset = 1) { setPCH(HI_BYTE(reg.pc) + offset); }
